+++ src/components/OwnBeats.tsx
@@ -1,10 +1,10 @@
 import React, { useEffect, useRef, useState } from "react";
 import type { SyntheticEvent } from "react";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import "../data/fa-library";
+import "~/lib/fa-library";
 import { api } from "~/utils/api";
 import type { Beat } from "@prisma/client";
-import type { Loop } from "~/data/loop";
+import type { Loop } from "~/types/loop";
 
 export default function OwnBeats({
     loadLoop,
