+++ src/containers/App.tsx
@@ -14,9 +14,9 @@ import {
 
 import MachineSequencer from "./MachineSequencer";
 import MachineKnobs from "./MachineKnobs";
-import { getObjectKeysUnsafe } from "~/data/helpers";
-import type { Track } from "~/data/tracks";
-import type { Loop } from "~/data/loop";
+import { getObjectKeysUnsafe } from "~/lib/helpers";
+import type { Track } from "~/types/tracks";
+import type { Loop } from "~/types/loop";
 
 interface Engine {
     gainNodes: Record<Track, GainNode> | undefined;
